+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Changed
+
+- [#27] Re-writing PHPUnit tests
+
 ## [v2.0.0] - 2019-03-26
 
 ### Changed
@@ -146,3 +150,5 @@ Composer.
 [#21]: https://github.com/jawira/case-converter/pull/21
 
 [v2.0.0]: https://github.com/jawira/case-converter/compare/v1.2.0...v2.0.0
+
+[#27]: https://github.com/jawira/case-converter/pull/27